move rule creation to utility
